requirements.txt
pandas>=2.0.0
twilio>=8.10.0
supabase>=2.0.0
python-dateutil>=2.8.0
python-dotenv>=1.0.0
